requirements.txt
# Core Dependencies
notion-client==2.2.1
openai==1.12.0
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.31.0
Pillow==10.2.0

# Logging and utilities
colorlog==6.8.2

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
